+++ src/Exceptions/InstagramAuthenticationException.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace CodebarAg\LaravelInstagram\Exceptions;
+
+class InstagramAuthenticationException extends InstagramException {}