+++ src/Exceptions/InstagramException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace CodebarAg\LaravelInstagram\Exceptions;
+
+use Exception;
+
+class InstagramException extends Exception {}